Fix file name
Seperate modules with client_class dir
